jobseeker CRUD feature complete

+++ Project_backend/src/main/java/com/matchmaking/backend/entity/Account.java
@@ -11,7 +11,7 @@ public class Account {
 
     private String password;
 
-    private Long accountId;
+    private int accountId;
 
     private int roleId;
